smc_signal_system/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dateutil>=2.8.0
requests>=2.31.0
tqdm>=4.66.0
pyarrow>=12.0.0
pytest>=7.4.0







